Split ntuple.py to multiple files
